backend/requirements.txt
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymongo==4.6.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
celery==5.3.1
redis==4.6.0
django-extensions==3.2.3
